refactor: Complete output refactor

+++ pkg/binman.go
@@ -1,6 +1,7 @@
 package binman
 
 import (
+	"fmt"
 	"os"
 	"reflect"
 	"runtime"
@@ -30,9 +31,10 @@ func goSyncRepo(ghClient *github.Client, rel BinmanRelease, c chan<- BinmanMsg, 
 		err = task.execute()
 		// this is hacky, but catches error
 		if err != nil && err.Error() == "Noupdate" {
+			c <- BinmanMsg{rel: rel, err: err}
 			return
 		} else if err != nil {
-			log.Warnf("Unable to complete action %s : %v", reflect.TypeOf(task), err)
+			log.Debugf("Unable to complete action %s : %v", reflect.TypeOf(task), err)
 			c <- BinmanMsg{rel: rel, err: err}
 			return
 		}
@@ -44,7 +46,7 @@ func goSyncRepo(ghClient *github.Client, rel BinmanRelease, c chan<- BinmanMsg, 
 		log.Debugf("Running task %s for %s", reflect.TypeOf(action), rel.Repo)
 		err = action.execute()
 		if err != nil {
-			log.Warnf("Unable to complete task %s : %v", reflect.TypeOf(action), err)
+			log.Debugf("Unable to complete task %s : %v", reflect.TypeOf(action), err)
 			c <- BinmanMsg{rel: rel, err: err}
 			return
 		}
@@ -54,8 +56,6 @@ func goSyncRepo(ghClient *github.Client, rel BinmanRelease, c chan<- BinmanMsg, 
 }
 
 func BinmanGetReleasePrep(work map[string]string) []BinmanRelease {
-
-	log.Infof("direct repo download")
 
 	if f, err := os.Stat(work["path"]); !f.IsDir() || err != nil {
 		log.Fatalf("Issue detected with %s", work["path"])
@@ -83,13 +83,21 @@ func BinmanGetReleasePrep(work map[string]string) []BinmanRelease {
 }
 
 // Main does basic setup, then calls the appropriate functions for asset resolution
-func Main(args map[string]string, debug bool, jsonLog bool, launchCommand string) {
+func Main(args map[string]string, debug bool, jsonLog bool, table bool, launchCommand string) {
+
+	spin, err := getSpinner(debug)
+	if err != nil {
+		log.Fatalf("Unable to get spinner - %s", err)
+	}
+
+	spin.Message("Binman sync begin")
 
 	// Set the logging options
 	log.ConfigureLog(jsonLog, debug)
-	log.Infof("binman sync begin")
+	log.Debugf("binman sync begin\n")
 
 	c := make(chan BinmanMsg)
+	output := make(map[string][]BinmanMsg)
 	var wg sync.WaitGroup
 	var releases []BinmanRelease
 	var ghClient *github.Client
@@ -111,11 +119,13 @@ func Main(args map[string]string, debug bool, jsonLog bool, launchCommand string
 		releases = config.Releases
 	}
 
-	log.Debugf("Process %v Releases", len(releases))
+	relLength := len(releases)
+	log.Debugf("Process %v Releases", relLength)
 
 	// https://github.com/lotusirous/go-concurrency-patterns/blob/main/2-chan/main.go
-	for _, rel := range releases {
+	for index, rel := range releases {
 		wg.Add(1)
+		setMessage(spin, fmt.Sprintf("Processing %d/%d %s", index+1, relLength, rel.Repo), 100)
 		go goSyncRepo(ghClient, rel, c, &wg)
 	}
 
@@ -126,20 +136,42 @@ func Main(args map[string]string, debug bool, jsonLog bool, launchCommand string
 
 	// Process results
 	for msg := range c {
+
+		setMessage(spin, "Downloading new releases", 0)
+
 		if msg.err == nil {
+			setMessage(spin, fmt.Sprintf("Downloaded %s ✅", msg.rel.Repo), 100)
+			output["Synced"] = append(output["Synced"], msg)
 			continue
 		}
 
-		log.Warnf("Repo %s, Error %q\n", msg.rel.Repo, msg.err)
+		if msg.err.Error() == "Noupdate" {
+			output["Up to Date"] = append(output["Up to Date"], msg)
+			continue
+		}
+
+		output["Error"] = append(output["Error"], msg)
 		if msg.rel.cleanupOnFailure {
 			err := os.RemoveAll(msg.rel.publishPath)
 			if err != nil {
-				log.Warnf("Unable to clean up %s - %s", msg.rel.publishPath, err)
+				log.Debugf("Unable to clean up %s - %s", msg.rel.publishPath, err)
 			}
-			log.Warnf("cleaned %s\n", msg.rel.publishPath)
+			log.Debugf("cleaned %s\n", msg.rel.publishPath)
 			log.Debugf("Final release data  %+v\n", msg.rel)
 		}
 	}
 
-	log.Infof("binman finished!")
+	spin.Suffix("")
+	spin.StopMessage(setStopMessage(output))
+	spin.Stop()
+
+	if table {
+		OutputResults(output, debug)
+	}
+
+	for _, msg := range output["Error"] {
+		fmt.Printf("%s : error = %v\n", msg.rel.Repo, msg.err)
+	}
+
+	log.Debugf("binman finished!")
 }
